Fixed C++ and python examples
