+++ examples/ar-advanced/rs-ar-advanced.cpp
@@ -9,6 +9,7 @@
 #include <fstream>
 #include <vector>
 #include "example.hpp"
+#include "../examples/example-utils.hpp"
 
 struct point3d {
     float f[3];
