+++ examples/measure/rs-measure.cpp
@@ -14,6 +14,7 @@
 #include <thread>
 #include <atomic>
 #include <mutex>
+#include "../examples/example-utils.hpp"
 
 using pixel = std::pair<int, int>;
