+++ examples/trajectory/rs-trajectory.cpp
@@ -4,6 +4,7 @@
 #include <librealsense2/rs.hpp>
 #include "example.hpp"          // Include short list of convenience functions for rendering
 #include <cstring>
+#include "../examples/example-utils.hpp"
 
 struct short3
 {
