+++ wrappers/python/examples/opencv_pointcloud_viewer.py
@@ -30,7 +30,6 @@ import time
 import cv2
 import numpy as np
 import pyrealsense2 as rs
-context = rs.context()
 
 class AppState:
